+++ History.md
@@ -1,3 +1,8 @@
+Unpublished
+===========
+
+  * allow sampling interval to be configured
+
 0.0.0 / Feb 05 2017
 ===================